+++ src/Env/Scanner/Scanner_Type.php
@@ -1,0 +1,10 @@
+<?php 
+
+namespace kranack\Lint\Env\Scanner;
+
+abstract class Scanner_Type
+{
+	const LOCAL = 'Local';
+	const MACPORTS = 'Macports';
+	const HOMEBREW = 'Homebrew';
+}